Add quiz categories (#29)
A quiz must of course have some kind of category!

This change will fix this by adding a category as a new field to the quiz schema. Initially I went for a new DB tables but realized that was not neccessary here.

This will also change the "add new quiz" form to have the ability to chose a category for the quiz.