Clean up unused code (#30)
Because I used the generators to generate schemas etc, a lot of unused code and tests was added to the codebase. This will clean up some some stuff that was not being used (tests for unused code is pretty useless I'd say). I also fixed some tests that were failing because of changed schemas and migrations.