Identity registration and login (#36)
Not everyone has a (or wants to use their) Google account to sign in, which is why the app needs a regular email/password authentication method. This adds that possibility by allowing a user to create an account with email and password.
This also adds views for login & register pages.